Metadata_processed.csv
This Metadata was the one used to carry out the analysis. 
Age was not taken into account since all ages are within the same range and no changes when analysing the same data using this metadata were observed.

Regarding to Sex, it was also removed because some of the algorihtms could not run when using this variable (correlations within the varibales were found). Those who could run, any changes were observed.

The research team thinks it may be due to the low number of samples, taking into consideratian that is an age related syndrome. It is believed that if more samples were obained some differences would be detected at least between age groups.